+++ tests/testthat/test-group-by.r
@@ -146,3 +146,9 @@ test_that("select(group_by(.)) implicitely adds grouping variables (#170)", {
   expect_equal(names(res), c("vs", "mpg"))
 })
 
+test_that("grouped_df errors on empty vars (#398)",{
+  m <- mtcars %>% group_by(cyl)
+  m <- m[1:3,1:3] 
+  expect_error( m %>% do(mpg = mean(.$mpg)) )
+})
+